getEAP_TLS/requirements.txt
djangorestframework
qrcode
pillow
pycryptodome
python-decouple
mysqlclient